Checkpoint the split of machine deployments
