+++ internal/command/deploy/machines.go
@@ -2,7 +2,6 @@ package deploy
 
 import (
 	"context"
-	"errors"
 	"fmt"
 	"strconv"
 	"strings"
@@ -16,7 +15,6 @@ import (
 	"github.com/superfly/flyctl/client"
 	"github.com/superfly/flyctl/flaps"
 	"github.com/superfly/flyctl/gql"
-	"github.com/superfly/flyctl/helpers"
 	"github.com/superfly/flyctl/internal/appconfig"
 	"github.com/superfly/flyctl/internal/cmdutil"
 	machcmd "github.com/superfly/flyctl/internal/command/machine"
@@ -174,47 +172,6 @@ func NewMachineDeployment(ctx context.Context, args MachineDeploymentArgs) (Mach
 	return md, nil
 }
 
-func (md *machineDeployment) runReleaseCommand(ctx context.Context) error {
-	if len(md.releaseCommand) == 0 || md.restartOnly {
-		return nil
-	}
-	io := iostreams.FromContext(ctx)
-	fmt.Fprintf(io.ErrOut, "Running %s release_command: %s\n",
-		md.colorize.Bold(md.app.Name),
-		md.appConfig.Deploy.ReleaseCommand,
-	)
-	err := md.createOrUpdateReleaseCmdMachine(ctx)
-	if err != nil {
-		return fmt.Errorf("error running release_command machine: %w", err)
-	}
-	releaseCmdMachine := md.releaseCommandMachine.GetMachines()[0]
-	// FIXME: consolidate this wait stuff with deploy waits? Especially once we improve the outpu
-	err = releaseCmdMachine.WaitForState(ctx, api.MachineStateStarted, md.waitTimeout)
-	if err != nil {
-		return fmt.Errorf("error waiting for release_command machine %s to start: %w", releaseCmdMachine.Machine().ID, err)
-	}
-	err = releaseCmdMachine.WaitForState(ctx, api.MachineStateDestroyed, md.waitTimeout)
-	if err != nil {
-		return fmt.Errorf("error waiting for release_command machine %s to finish running: %w", releaseCmdMachine.Machine().ID, err)
-	}
-	lastExitEvent, err := releaseCmdMachine.WaitForEventTypeAfterType(ctx, "exit", "start", md.waitTimeout)
-	if err != nil {
-		return fmt.Errorf("error finding the release_command machine %s exit event: %w", releaseCmdMachine.Machine().ID, err)
-	}
-	exitCode, err := lastExitEvent.Request.GetExitCode()
-	if err != nil {
-		return fmt.Errorf("error get release_command machine %s exit code: %w", releaseCmdMachine.Machine().ID, err)
-	}
-	if exitCode != 0 {
-		fmt.Fprintf(md.io.ErrOut, "Error release_command failed running on machine %s with exit code %s. Check the logs at: https://fly.io/apps/%s/monitoring\n",
-			md.colorize.Bold(releaseCmdMachine.Machine().ID), md.colorize.Red(strconv.Itoa(exitCode)), md.app.Name)
-		return fmt.Errorf("error release_command machine %s exited with non-zero status of %d", releaseCmdMachine.Machine().ID, exitCode)
-	}
-	md.logClearLinesAbove(1)
-	fmt.Fprintf(md.io.ErrOut, "  release_command %s completed successfully\n", md.colorize.Bold(releaseCmdMachine.Machine().ID))
-	return nil
-}
-
 func (md *machineDeployment) resolveProcessGroupChanges() ProcessGroupsDiff {
 	output := ProcessGroupsDiff{
 		groupsToRemove:        map[string]int{},
@@ -328,7 +285,29 @@ func (md *machineDeployment) spawnMachineInGroup(ctx context.Context, groupName 
 
 func (md *machineDeployment) DeployMachinesApp(ctx context.Context) error {
 	ctx = flaps.NewContext(ctx, md.flapsClient)
+	if md.restartOnly {
+		return md.restartMachinesApp(ctx)
+	}
+	return md.deployMachinesApp(ctx)
+}
 
+func (md *machineDeployment) restartMachinesApp(ctx context.Context) error {
+	err := md.machineSet.AcquireLeases(ctx, md.leaseTimeout)
+	defer func() {
+		err := md.machineSet.ReleaseLeases(ctx)
+		if err != nil {
+			terminal.Warnf("error releasing leases on machines: %v\n", err)
+		}
+	}()
+	if err != nil {
+		return err
+	}
+	md.machineSet.StartBackgroundLeaseRefresh(ctx, md.leaseTimeout, md.leaseDelayBetween)
+
+	return md.updateExistingMachines(ctx)
+}
+
+func (md *machineDeployment) deployMachinesApp(ctx context.Context) error {
 	err := md.runReleaseCommand(ctx)
 	if err != nil {
 		return fmt.Errorf("release command failed - aborting deployment. %w", err)
@@ -361,18 +340,6 @@ func (md *machineDeployment) DeployMachinesApp(ctx context.Context) error {
 	md.machineSet.StartBackgroundLeaseRefresh(ctx, md.leaseTimeout, md.leaseDelayBetween)
 
 	processGroupMachineDiff := md.resolveProcessGroupChanges()
-
-	// If restartOnly is set, that means we're *re*deploying a configuration.
-	// It also probably means that we're in a context (like setting secrets) where
-	// creating or destroying machines would be unexpected.
-	if md.restartOnly {
-		removingMachines := len(processGroupMachineDiff.machinesToRemove) != 0
-		addingMachines := len(processGroupMachineDiff.groupsNeedingMachines) != 0
-		if removingMachines || addingMachines {
-			return errors.New("your app's machines don't match the remote configuration.\n[!] running 'fly deploy' would probably help")
-		}
-	}
-
 	md.warnAboutProcessGroupChanges(ctx, processGroupMachineDiff)
 
 	// Destroy machines that don't fit the current process groups
@@ -395,8 +362,11 @@ func (md *machineDeployment) DeployMachinesApp(ctx context.Context) error {
 		fmt.Fprintf(md.io.ErrOut, "Finished launching new machines\n")
 	}
 
-	// FIXME: handle deploy strategy: rolling, immediate, canary, bluegreen
+	return md.updateExistingMachines(ctx)
+}
 
+func (md *machineDeployment) updateExistingMachines(ctx context.Context) error {
+	// FIXME: handle deploy strategy: rolling, immediate, canary, bluegreen
 	fmt.Fprintf(md.io.Out, "Updating existing machines in '%s' with %s strategy\n", md.colorize.Bold(md.app.Name), md.strategy)
 	for _, m := range md.machineSet.GetMachines() {
 		launchInput := md.resolveUpdatedMachineConfig(m.Machine(), false)
@@ -469,98 +439,6 @@ func (md *machineDeployment) setMachinesForDeployment(ctx context.Context) error
 		releaseCmdSet = []*api.Machine{releaseCmdMachine}
 	}
 	md.releaseCommandMachine = machine.NewMachineSet(md.flapsClient, md.io, releaseCmdSet)
-	return nil
-}
-
-func (md *machineDeployment) createOrUpdateReleaseCmdMachine(ctx context.Context) error {
-	if md.releaseCommandMachine.IsEmpty() {
-		return md.createReleaseCommandMachine(ctx)
-	} else {
-		return md.updateReleaseCommandMachine(ctx)
-	}
-}
-
-func (md *machineDeployment) configureLaunchInputForReleaseCommand(launchInput *api.LaunchMachineInput) *api.LaunchMachineInput {
-	launchInput.Config.Init.Cmd = md.releaseCommand
-	launchInput.Config.Metadata[api.MachineConfigMetadataKeyFlyProcessGroup] = api.MachineProcessGroupFlyAppReleaseCommand
-	launchInput.Config.Restart = api.MachineRestart{
-		Policy: api.MachineRestartPolicyNo,
-	}
-	launchInput.Config.AutoDestroy = true
-	launchInput.Config.DNS = &api.DNSConfig{SkipRegistration: true}
-	if md.appConfig.PrimaryRegion != "" {
-		launchInput.Region = md.appConfig.PrimaryRegion
-	}
-	if _, present := launchInput.Config.Env["RELEASE_COMMAND"]; !present {
-		launchInput.Config.Env["RELEASE_COMMAND"] = "1"
-	}
-	if launchInput.Config.Guest == nil {
-		launchInput.Config.Guest = &api.MachineGuest{}
-	}
-
-	desiredGuest := helpers.Clone(api.MachinePresets["shared-cpu-2x"])
-	if !md.machineSet.IsEmpty() {
-		group := md.appConfig.DefaultProcessName()
-		ram := func(m *api.Machine) int {
-			if m != nil && m.Config != nil && m.Config.Guest != nil {
-				return m.Config.Guest.MemoryMB
-			}
-			return 0
-		}
-		maxRamMach := lo.Reduce(md.machineSet.GetMachines(), func(prevBest *api.Machine, lm machine.LeasableMachine, _ int) *api.Machine {
-			mach := lm.Machine()
-			if mach.ProcessGroup() != group {
-				return prevBest
-			}
-			return lo.Ternary(ram(mach) > ram(prevBest), mach, prevBest)
-		}, nil)
-		if maxRamMach != nil {
-			desiredGuest = maxRamMach.Config.Guest
-		}
-	}
-	launchInput.Config.Guest = helpers.Clone(desiredGuest)
-	return launchInput
-}
-
-func (md *machineDeployment) createReleaseCommandMachine(ctx context.Context) error {
-	if len(md.releaseCommand) == 0 || !md.releaseCommandMachine.IsEmpty() {
-		return nil
-	}
-	launchInput := md.resolveUpdatedMachineConfig(nil, true)
-	releaseCmdMachine, err := md.flapsClient.Launch(ctx, *launchInput)
-	if err != nil {
-		return fmt.Errorf("error creating a release_command machine: %w", err)
-	}
-	fmt.Fprintf(md.io.ErrOut, "  Created release_command machine %s\n", md.colorize.Bold(releaseCmdMachine.ID))
-	md.releaseCommandMachine = machine.NewMachineSet(md.flapsClient, md.io, []*api.Machine{releaseCmdMachine})
-	return nil
-}
-
-func (md *machineDeployment) updateReleaseCommandMachine(ctx context.Context) error {
-	if len(md.releaseCommand) == 0 {
-		return nil
-	}
-	if md.releaseCommandMachine.IsEmpty() {
-		return fmt.Errorf("expected release_command machine to exist already, but it does not :-(")
-	}
-	releaseCmdMachine := md.releaseCommandMachine.GetMachines()[0]
-	fmt.Fprintf(md.io.ErrOut, "  Updating release_command machine %s\n", md.colorize.Bold(releaseCmdMachine.Machine().ID))
-	err := releaseCmdMachine.WaitForState(ctx, api.MachineStateStopped, md.waitTimeout)
-	if err != nil {
-		return err
-	}
-	updatedConfig := md.resolveUpdatedMachineConfig(releaseCmdMachine.Machine(), true)
-	err = md.releaseCommandMachine.AcquireLeases(ctx, md.leaseTimeout)
-	defer func() {
-		_ = md.releaseCommandMachine.ReleaseLeases(ctx)
-	}()
-	if err != nil {
-		return err
-	}
-	err = releaseCmdMachine.Update(ctx, *updatedConfig)
-	if err != nil {
-		return fmt.Errorf("error updating release_command machine: %w", err)
-	}
 	return nil
 }
 
@@ -685,7 +563,36 @@ func (md *machineDeployment) createReleaseInBackend(ctx context.Context) error {
 	return nil
 }
 
+func (md *machineDeployment) computeMachineConfigMetadata(config *api.MachineConfig) map[string]string {
+	return lo.Assign(
+		md.defaultMachineMetadata(),
+		lo.OmitBy(config.Metadata, func(k, v string) bool {
+			return isFlyAppsPlatformMetadata(k)
+		}))
+}
+
+func (md *machineDeployment) launchInputForRestart(origMachineRaw *api.Machine) *api.LaunchMachineInput {
+	if origMachineRaw == nil {
+		return nil
+	}
+
+	Config := machine.CloneConfig(origMachineRaw.Config)
+	Config.Metadata = md.computeMachineConfigMetadata(Config)
+
+	return &api.LaunchMachineInput{
+		ID:      origMachineRaw.ID,
+		AppID:   md.app.Name,
+		OrgSlug: md.app.Organization.ID,
+		Config:  Config,
+		Region:  origMachineRaw.Region,
+	}
+}
+
 func (md *machineDeployment) resolveUpdatedMachineConfig(origMachineRaw *api.Machine, forReleaseCommand bool) *api.LaunchMachineInput {
+	if md.restartOnly {
+		return md.launchInputForRestart(origMachineRaw)
+	}
+
 	if origMachineRaw == nil {
 		origMachineRaw = &api.Machine{
 			Region: md.appConfig.PrimaryRegion,
@@ -701,22 +608,7 @@ func (md *machineDeployment) resolveUpdatedMachineConfig(origMachineRaw *api.Mac
 		Region:  origMachineRaw.Region,
 	}
 
-	if launchInput.Config.Metadata == nil {
-		launchInput.Config.Metadata = map[string]string{}
-	}
-
-	launchInput.Config.Metadata = lo.Assign(
-		md.defaultMachineMetadata(),
-		lo.OmitBy(launchInput.Config.Metadata, func(k, v string) bool {
-			return isFlyAppsPlatformMetadata(k)
-		}),
-	)
-
-	// Stop here If the machine is restarting
-	if md.restartOnly {
-		return launchInput
-	}
-
+	launchInput.Config.Metadata = md.computeMachineConfigMetadata(launchInput.Config)
 	launchInput.Config.Statics = nil
 	launchInput.Config.Image = md.img
 	launchInput.Config.Env = lo.Assign(md.appConfig.Env)
